requirements.txt
streamlit>=1.24.0
firebase-admin>=6.2.0
plotly>=5.15.0
pyrebase4>=4.7.1
python-dotenv>=1.0.0
python-jwt>=4.0.0
requests>=2.31.0
google-cloud-storage>=2.14.0
Pillow>=10.0.0; platform_python_implementation != 'PyPy'
protobuf>=4.21.0
setuptools>=68.0.0
wheel>=0.40.0
gcloud>=0.18.3
oauth2client>=4.1.3
pandas>=2.0.0
openpyxl>=3.1.0
